tests: separe track tests and improve before each
